rename supported-login-methods to auth-form-helpers and delete old supported-auth-backends helper, update tests
