+++ ui/.template-lintrc.js
@@ -10,9 +10,6 @@ module.exports = {
 
   rules: {
     'no-action': 'off',
-    'no-implicit-this': {
-      allow: ['supported-auth-backends'],
-    },
     'require-input-label': 'off',
     'no-array-prototype-extensions': 'off',
     // from bump to ember-template-lint@6.0.0
