+++ ui/tests/integration/components/ldap/accounts-checked-out-test.js
@@ -57,6 +57,10 @@ module('Integration | Component | ldap | AccountsCheckedOut', function (hooks) {
     };
   });
 
+  hooks.afterEach(function () {
+    this.authStub.restore();
+  });
+
   test('it should render empty state when no accounts are checked out', async function (assert) {
     this.statuses = [
       { account: 'foo', available: true, library: 'test-library' },
@@ -74,7 +78,7 @@ module('Integration | Component | ldap | AccountsCheckedOut', function (hooks) {
   });
 
   test('it should filter accounts for root user', async function (assert) {
-    this.authStub.value({});
+    this.authStub.value({ entityId: '' });
 
     await this.renderComponent();
 
@@ -85,7 +89,7 @@ module('Integration | Component | ldap | AccountsCheckedOut', function (hooks) {
   });
 
   test('it should filter accounts for non root user', async function (assert) {
-    this.authStub.value({ entity_id: '456' });
+    this.authStub.value({ entityId: '456' });
 
     await this.renderComponent();
 
@@ -107,7 +111,7 @@ module('Integration | Component | ldap | AccountsCheckedOut', function (hooks) {
   });
 
   test('it should display details in table', async function (assert) {
-    this.authStub.value({ entity_id: '456' });
+    this.authStub.value({ entityId: '456' });
 
     await this.renderComponent();
